services/actions-tokenize/requirements.txt
acryl-datahub>=0.12.1.4
psycopg2-binary>=2.9
pydantic>=1.10
